feat: Add biner button adn slider function
Add slider biner function. The binary button contains a process to change the grayscale result to binary according to the position value on the trackbar.
If the grayscale value <= trackbar value, the pixels are set to black or RGB (0.0,0). If grayscale value> trackbar value, then the pixels are set to white or RGB (255,255,255).